Extract FlatpakContext to a separate file
This is basically a code motion only, no changes to behaviour.
