+++ common/flatpak-common-types.h
@@ -28,7 +28,6 @@ typedef enum {
 
 typedef struct FlatpakDir     FlatpakDir;
 typedef struct FlatpakDeploy  FlatpakDeploy;
-typedef struct FlatpakContext FlatpakContext;
 typedef struct FlatpakOciRegistry FlatpakOciRegistry;
 typedef struct _FlatpakOciManifest FlatpakOciManifest;
